pyproject.toml
[project]
name = "google-docs-mcp"
version = "1.0.0"
description = "MCP server for native Google Docs editing in Claude Cowork"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Sasha Kang", email = "sasha.kang@gmail.com" }]
keywords = ["mcp", "google-docs", "claude", "cowork"]

dependencies = [
    "mcp>=1.0.0,<2.0.0",
    "google-auth>=2.23.0,<3.0.0",
    "google-auth-oauthlib>=1.1.0,<2.0.0",
    "google-api-python-client>=2.100.0,<3.0.0",
    "tenacity>=8.2.0,<9.0.0",
]

[project.urls]
Repository = "https://github.com/sashakang/google-docs-mcp"
Issues = "https://github.com/sashakang/google-docs-mcp/issues"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["server*"]
